Unificacion de componentes
